requirements.txt
python-telegram-bot
VaderSentiment
openai
regex
python-dotenv
pyttsx3
